refactor: share messaging dashboard component
